Refactor: transaction calculations on server
