+++ backend/src/procedure-handlers.ts
@@ -1,7 +1,7 @@
 import { exec } from "child_process";
 import { promisify } from "util";
 
-import { KeysOfUnion, ProcedureTypes, RPC } from "./types";
+import { Action, KeysOfUnion, ProcedureTypes, RPC } from "./types";
 
 import * as stats from "./stats";
 import * as receipts from "./receipts";
@@ -27,13 +27,6 @@ export const procedureHandlers: {
 } = {
   "node-telemetry": async ([nodeInfo]) => {
     return await telemetry.sendTelemetry(nodeInfo);
-  },
-
-  "nearcore-tx": async ([transactionHash, accountId]) => {
-    return await sendJsonRpc("EXPERIMENTAL_tx_status", [
-      transactionHash,
-      accountId,
-    ]);
   },
 
   "nearcore-final-block": async () => {
@@ -109,7 +102,80 @@ export const procedureHandlers: {
   },
 
   "transaction-info": async ([transactionHash]) => {
-    return await transactions.getTransactionInfo(transactionHash);
+    const transactionBaseInfo = await transactions.getTransactionInfo(
+      transactionHash
+    );
+    if (!transactionBaseInfo) {
+      return null;
+    }
+    const transactionInfo = await sendJsonRpc("EXPERIMENTAL_tx_status", [
+      transactionBaseInfo.hash,
+      transactionBaseInfo.signerId,
+    ]);
+
+    const actions = transactionInfo.transaction.actions.map(
+      transactions.mapRpcActionToAction
+    );
+    const receipts = transactionInfo.receipts;
+    const receiptsOutcome = transactionInfo.receipts_outcome;
+    if (
+      receipts.length === 0 ||
+      receipts[0].receipt_id !== receiptsOutcome[0].id
+    ) {
+      receipts.unshift({
+        predecessor_id: transactionInfo.transaction.signer_id,
+        receipt: {
+          Action: {
+            signer_id: transactionInfo.transaction.signer_id,
+            signer_public_key: "",
+            gas_price: "0",
+            output_data_receivers: [],
+            input_data_ids: [],
+            actions: transactionInfo.transaction.actions,
+          },
+        },
+        receipt_id: receiptsOutcome[0].id,
+        receiver_id: transactionInfo.transaction.receiver_id,
+      });
+    }
+    const receiptOutcomesByIdMap = new Map<
+      string,
+      RPC.ExecutionOutcomeWithIdView
+    >();
+    receiptsOutcome.forEach((receipt) => {
+      receiptOutcomesByIdMap.set(receipt.id, receipt);
+    });
+
+    const receiptsByIdMap = new Map<
+      string,
+      Omit<RPC.ReceiptView, "actions"> & { actions: Action[] }
+    >();
+    receipts.forEach((receiptItem) => {
+      receiptsByIdMap.set(receiptItem.receipt_id, {
+        ...receiptItem,
+        actions:
+          "Action" in receiptItem.receipt
+            ? receiptItem.receipt.Action.actions.map(
+                transactions.mapRpcActionToAction
+              )
+            : [],
+      });
+    });
+
+    return {
+      ...transactionBaseInfo,
+      status: Object.keys(
+        transactionInfo.status
+      )[0] as KeysOfUnion<RPC.FinalExecutionStatus>,
+      actions,
+      receiptsOutcome,
+      receipt: transactions.collectNestedReceiptWithOutcome(
+        receiptsOutcome[0].id,
+        receiptsByIdMap,
+        receiptOutcomesByIdMap
+      ),
+      transactionOutcome: transactionInfo.transaction_outcome,
+    };
   },
 
   "transaction-execution-status": async ([hash, signerId]) => {
