+++ frontend/src/components/transactions/TransactionDetails.tsx
@@ -14,9 +14,8 @@ import TransactionExecutionStatus from "./TransactionExecutionStatus";
 
 import { useTranslation } from "react-i18next";
 import { useFinalBlockTimestampNanosecond } from "../../hooks/data";
-import { Transaction } from "../../pages/transactions/[hash]";
 import { styled } from "../../libraries/styles";
-import { RPC } from "../../types/common";
+import { RPC, Transaction } from "../../types/common";
 
 const HeaderRow = styled(Row);
